Rename cluster controller to kubernetes controller
