+++ api/pkg/controller/seed-controller-manager/kubernetes/cluster_controller.go
@@ -1,4 +1,4 @@
-package cluster
+package kubernetes
 
 import (
 	"context"
