+++ api/pkg/controller/seed-controller-manager/kubernetes/resources.go
@@ -1,4 +1,4 @@
-package cluster
+package kubernetes
 
 import (
 	"context"
